pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "foundry-agent-studio"
version = "0.1.0"
description = "Local AI player characters for Foundry VTT (Ollama, SQLite, HTTP bridge)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "uvicorn[standard]>=0.32.0",
  "piper-tts>=1.4.0,<2",
]

[project.scripts]
fas = "foundry_agent_studio.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["foundry_agent_studio*"]

[tool.setuptools.package-data]
foundry_agent_studio = [
  "data/piper_voice_catalog.txt",
  "voices/*.onnx",
  "voices/*.onnx.json",
  "voices/README.txt",
  "whisper_models/README.txt",
  "bin/windows/*.dll",
  "bin/windows/*.exe",
  "bin/windows/README.txt",
]
